Role mapper should check if an update is needed for the role
Closes #43698

+++ services/src/main/java/org/keycloak/broker/saml/mappers/AbstractAttributeToRoleMapper.java
@@ -62,9 +62,13 @@ public abstract class AbstractAttributeToRoleMapper extends AbstractIdentityProv
         if (!context.hasMapperGrantedRole(roleName)) {
             if (this.applies(mapperModel, context)) {
                 context.addMapperGrantedRole(roleName);
-                user.grantRole(role);
+                if (user.getRealmRoleMappingsStream().noneMatch(r -> r.equals(role))) {
+                    user.grantRole(role);
+                }
             } else {
-                user.deleteRoleMapping(role);
+                if (user.getRealmRoleMappingsStream().anyMatch(r -> r.equals(role))) {
+                    user.deleteRoleMapping(role);
+                }
             }
         }
     }
